Improve detection of first party apps.
Django apps sometimes are referenced with their dotted path in `settings.FIRST_PARTY_APPS`, e.g. `myproject.myapp`. But the label of the app config is the right most part of the path, e.g. `myapp` which is not in `settings.FIRST_PARTY_APPS` and thus the app will not be considered a first party app.

We can improve this by looking at the name of the app config, which should normally correspond to the dotted path of the app.

A similar problem arises if `settings.FIRST_PARTY_APPS` contains the dotted path to the app's app config class, e.g. `myproject.myapp.apps.MyAppConfigSubclass`. But that can be worked around by putting `default_app_config` into the apps `__init__.py`, so we are not addressing that right now.