rebase-migration: Make it possible to handle chains of migrations.
Previously, the rebase_migration command was not able to rebase
chains of migrations in a single app.

This commit introduces a new flag -- "new" which basically is used
for the first migration you create, and it wipes out your migration
history in max_migration.txt, and writes up that first migration in
the first line. Any further migrations added without the flag are
simply added under each other in the max_migration.txt.

This would allow the rebase_migration command to access the
chain of migrations that need to be rebased in a commit, and
will rebase them accordingly.

Fixes #27