rebase_migration: Unapply migrations while rebase.
Previously, while rebasing migrations, if the migration to be
rebased is already applied, it would throw out an error asking
to un apply the migration before rebasing.

This commit un applies the migrations and rebases them when
the migrations to be rebased are already applied. This is done
by first creating a --merge migration, un applying the previous
migrations, removing the --merge migration, and then rebasing
the existing migrations.

Fixes #184